Added Meta information option
The meta information is displayed under the main Tweet. It uses jQuery
to initially hide the element - ensuring that if javascript is disabled
the user will still be able to see the meta information. The meta can be
shown or hidden (on a toggle) by clicking the '...' span, which has a
class of breaker.